resolved bug

+++ inc/philo.h
@@ -6,7 +6,7 @@
 /*   By: doduwole <doduwole@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/05 19:44:36 by doduwole          #+#    #+#             */
-/*   Updated: 2023/09/29 10:31:05 by doduwole         ###   ########.fr       */
+/*   Updated: 2023/09/29 13:00:42 by doduwole         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,18 +43,18 @@ typedef enum e_args
 
 typedef struct s_time
 {
-	uint64_t				to_die;
-	uint64_t				to_eat;
-	uint64_t				to_sleep;
+	int				to_die;
+	int				to_eat;
+	int				to_sleep;
 }					t_time;
 
 typedef struct s_data
 {
 	int				philo_num;
 	t_time			time;
-	uint64_t		n_eat;
-	uint64_t		philo_died;
-	uint64_t		timer;
+	int		n_eat;
+	int		philo_died;
+	long long		timer;
 	pthread_t		monitor;
 	t_mutex			*mymutex;
 	t_mutex			shared;
@@ -70,9 +70,9 @@ typedef struct s_fork
 
 typedef struct s_philo
 {
-	uint64_t		phi_id;
-	uint64_t		t_die;
-	uint64_t		n_eaten;
+	int		phi_id;
+	long long		t_die;
+	int		n_eaten;
 	t_fork			hand;
 	pthread_t		th;
 	t_data			*data;
